+++ lib/column_api.rb
@@ -16,6 +16,7 @@ module ColumnApi
 
   # Objects
   autoload :AchTransfer, "column_api/objects/ach_transfer"
+  autoload :AchReturn, "column_api/objects/ach_return"
   autoload :AccountNumber, "column_api/objects/account_number"
   autoload :BankAccount, "column_api/objects/bank_account"
   autoload :BusinessEntity, "column_api/objects/business_entity"
@@ -24,6 +25,7 @@ module ColumnApi
 
   # Resources
   autoload :AchTransferResource, "column_api/resources/ach_transfer_resource"
+  autoload :AchReturnResource, "column_api/resources/ach_return_resource"
   autoload :AccountNumberResource, "column_api/resources/account_number_resource"
   autoload :BankAccountResource, "column_api/resources/bank_account_resource"
   autoload :EntityResource, "column_api/resources/entity_resource"